refactor: Article 도메인과 Comment도메인 분리
- 기존에 Article 도메인에 있던 Comment관련 로직을 분리했습니다.
- 게시글 수정 로직에서 게시판 수정이 제대로 되지 않던 문제 해결
