+++ src/main/kotlin/com/arabyte/arabyteapi/domain/article/dto/ArticleLikeRequest.kt
@@ -1,4 +1,4 @@
-package com.arabyte.arabyteapi.domain.article.dto.article
+package com.arabyte.arabyteapi.domain.article.dto
 
 data class ArticleLikeRequest(
     val articleId: Long,
