+++ src/main/kotlin/com/arabyte/arabyteapi/domain/article/dto/CreateArticleRequest.kt
@@ -1,4 +1,4 @@
-package com.arabyte.arabyteapi.domain.article.dto.article
+package com.arabyte.arabyteapi.domain.article.dto
 
 import com.arabyte.arabyteapi.domain.article.enums.ArticleKind
