+++ src/main/kotlin/com/arabyte/arabyteapi/domain/article/dto/DeleteArticleResponse.kt
@@ -1,4 +1,4 @@
-package com.arabyte.arabyteapi.domain.article.dto.article
+package com.arabyte.arabyteapi.domain.article.dto
 
 data class DeleteArticleResponse(
     val articleId: Long,
