+++ src/main/kotlin/com/arabyte/arabyteapi/domain/article/dto/UpdateArticleRequest.kt
@@ -1,0 +1,10 @@
+package com.arabyte.arabyteapi.domain.article.dto
+
+import com.arabyte.arabyteapi.domain.article.enums.ArticleKind
+
+data class UpdateArticleRequest(
+    val title: String,
+    val text: String,
+    val isAnonymous: Boolean,
+    val articleKind: ArticleKind
+)
