--- src/main/kotlin/com/arabyte/arabyteapi/domain/article/dto/article/UpdateArticleRequest.kt
@@ -1,7 +1,0 @@
-package com.arabyte.arabyteapi.domain.article.dto.article
-
-data class UpdateArticleRequest(
-    val title: String,
-    val text: String,
-    val isAnonymous: Boolean
-)
