+++ src/main/kotlin/com/arabyte/arabyteapi/domain/article/service/ArticleService.kt
@@ -1,11 +1,12 @@
 package com.arabyte.arabyteapi.domain.article.service
 
+import com.arabyte.arabyteapi.domain.article.dto.*
 import com.arabyte.arabyteapi.domain.article.dto.article.*
-import com.arabyte.arabyteapi.domain.article.dto.comment.CommentResponse
+import com.arabyte.arabyteapi.domain.comment.dto.CommentResponse
 import com.arabyte.arabyteapi.domain.article.entity.Article
 import com.arabyte.arabyteapi.domain.article.enums.ArticleKind
 import com.arabyte.arabyteapi.domain.article.repository.ArticleRepository
-import com.arabyte.arabyteapi.domain.article.repository.CommentRepository
+import com.arabyte.arabyteapi.domain.comment.repository.CommentRepository
 import com.arabyte.arabyteapi.domain.user.service.UserService
 import com.arabyte.arabyteapi.global.enums.CustomError
 import com.arabyte.arabyteapi.global.exception.CustomException
@@ -122,6 +123,7 @@ class ArticleService(
         article.title = request.title
         article.text = request.text
         article.isAnonymous = request.isAnonymous
+        article.articleKindId = request.articleKind
 
         articleRepository.save(article)
