+++ src/main/kotlin/com/arabyte/arabyteapi/domain/comment/dto/CommentResponse.kt
@@ -1,4 +1,4 @@
-package com.arabyte.arabyteapi.domain.article.dto.comment
+package com.arabyte.arabyteapi.domain.comment.dto
 
 data class CommentResponse(
     val commentId: Long,
