+++ src/main/kotlin/com/arabyte/arabyteapi/domain/comment/dto/CreateCommentRequest.kt
@@ -1,9 +1,9 @@
-package com.arabyte.arabyteapi.domain.article.dto.comment
+package com.arabyte.arabyteapi.domain.comment.dto
 
 data class CreateCommentRequest(
     val articleId: Long,
     val userId: Long,
     val text: String,
-    val parentId: Long? = null,
+    val parentId: Long?,
     val isAnonymous: Boolean
 )
