+++ src/main/kotlin/com/arabyte/arabyteapi/domain/comment/repository/CommentRepository.kt
@@ -1,6 +1,6 @@
-package com.arabyte.arabyteapi.domain.article.repository
+package com.arabyte.arabyteapi.domain.comment.repository
 
-import com.arabyte.arabyteapi.domain.article.entity.Comment
+import com.arabyte.arabyteapi.domain.comment.entity.Comment
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
